requirements.txt
discord.py==1.5.1
aiofiles==0.5.0
aiohttp==3.6.2
mongoengine==0.20.0
marshmallow==3.7.1